+++ SMA/src/pk/lums/edu/sma/processing/EntityProcessor.java
@@ -39,7 +39,7 @@ public class EntityProcessor {
 	String csvEntities = entityLine[0].substring(1);
 	csvEntities = csvEntities.substring(0, csvEntities.length() - 1);
 	String[] entityArr = csvEntities.split(",");
-	Map<String, Integer> entityMap = new HashMap<String, Integer>();
+	Map<String, Double> entityMap = new HashMap<String, Double>();
 	for (String entity : entityArr) {
 	    if (entity.length() > 3) {
 		// System.out.println(entity);
@@ -47,7 +47,7 @@ public class EntityProcessor {
 		if (keyVal.length == 2 && keyVal[0].trim().length() > 4) {
 		    String key = keyVal[0].trim();
 		    int val = Integer.parseInt(keyVal[1]);
-		    entityMap.put(key, val);
+		    entityMap.put(key, (double) val);
 		}
 	    }
 	}
